linter

+++ libs/TestRtAudio/TestRtAudioLib/Components/RangeUtils.cpp
@@ -1,6 +1,6 @@
 #include "RangeUtils.h"
 
-float rescaleRange(float x, float in_min, float in_max, float out_min, float out_max)
+double rescaleRange(double x, double in_min, double in_max, double out_min, double out_max)
 {
   return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
@@ -12,7 +12,7 @@ double amplitudeFromDb(double amp)
   return pow(dbBase, amp / dbPow);
 }
 
-float constrainRange(float x, float min_val, float max_val)
+double constrainRange(double x, double min_val, double max_val)
 {
   if (x < min_val)
     return min_val;
