+++ libs/TestRtAudio/TestRtAudioLib/Components/RangeUtils.h
@@ -2,11 +2,11 @@
 #include <cmath>
 #include <vector>
 
-float rescaleRange(float x, float in_min, float in_max, float out_min, float out_max);
+double rescaleRange(double x, double in_min, double in_max, double out_min, double out_max);
 
-float constrainRange(float x, float min_val, float max_val);
+double constrainRange(double x, double min_val, double max_val);
 
-float midiNoteToFrequency(float noteNumber);
+double midiNoteToFrequency(double noteNumber);
 
 double amplitudeFromDb(double amp);
